requirements.txt

dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.3.1
nsepy==0.8
numpy==1.18.5
oauthlib==3.1.0
packaging==20.8
pandas==1.1.1
pandas-datareader==0.9.0
pandocfilters==1.4.3
Pillow==8.2.0
plotly==4.14.3
psycopg2==2.8.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.1
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
